Move elements to printer folder
